% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{avatar_palettes}
\alias{avatar_palettes}
\title{Complete list of Avatar palettes}
\format{
An object of class \code{list} of length 16.
}
\usage{
avatar_palettes
}
\description{
Use \code{\link{avatar_palette}} to construct palettes of desired length.
}
\keyword{datasets}
