% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avatar.R
\docType{package}
\name{AvatarTLA-package}
\alias{AvatarTLA}
\alias{AvatarTLA-package}
\title{AvatarTLA: Color Palettes Inspired by Avatar: The Last Airbender}
\description{
Color palettes inspired by characters and nations from the animated series Avatar: The Last Airbender for use in R graphics.
}
\details{
Color palettes inspired by characters and nations from the animated series Avatar: The Last Airbender for use in R graphics.
}
\author{
\strong{Maintainer}: Gerardo Manzanarez-Villasana \email{jupiter.gmv@gmail.com}

}
