% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{avatar_palette}
\alias{avatar_palette}
\title{An Avatar: The Last Airbender palette generator}
\usage{
avatar_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Mai}
\code{Suki}
\code{Iroh}
\code{FireNation}
\code{AirNomads}
\code{EarthKingdom}
\code{WaterTribe}
\code{Appa}
\code{Sokka}
\code{Toph}
\code{Azula}
\code{Aang}
\code{Zuko}
\code{Katara}}

\item{n}{Number of desired colors. If omitted, uses all colors from the palette.}

\item{type}{Either "discrete" or "continuous". Use continuous if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Color palettes inspired by characters and nations from Avatar: The Last Airbender.
}
\examples{
avatar_palette("Toph")
avatar_palette("Aang", 3)
avatar_palette("EarthKingdom", 30, type = "continuous")
}
\keyword{colors}
